requirements.txt
python-telegram-bot --pre
Pillow